renamed to thread_session.rb
