+++ lib/foreman/thread_session.rb
